neue Version

+++ BruttoRechner/BruttoRechner/BruttoRechner.cpp
@@ -3,16 +3,16 @@
 int main() {
 
 	float nettoBetrag = 0, bruttoBetrag = 0;
-	float steuersatz = 0;
+	int steuersatz = 0;
 
 	printf("---------------Brutto Rechner---------------\n\n");
 	printf("Geben Sie einen Netto Betrag an: ");
 	scanf("%f", &nettoBetrag);
 
 	printf("\nGeben Sie einen Steuersatz als Ganzzahl ein: ");
-	scanf("%f", &steuersatz);
+	scanf("%i", &steuersatz);
 
-	bruttoBetrag = nettoBetrag * (1 + steuersatz / 100);
+	bruttoBetrag = nettoBetrag * (1 + (float)steuersatz / 100);
 	printf("\nDer Bruttobetrag lautet: %.2f\n", bruttoBetrag);
